user adds charged particle by touching the screen.

+++ electric-android/app/src/main/java/com/github/fields/electric/Charge.java
@@ -3,14 +3,14 @@ package com.github.fields.electric;
 import android.graphics.Point;
 
 /**
+ * Charged particle.
  * Created by Moshe on 2017/01/28.
  */
-
 public class Charge extends Point {
 
-    public int size;
+    public double size;
 
-    public Charge(int x, int y, int size) {
+    public Charge(int x, int y, double size) {
         super(x, y);
         this.size = size;
     }
@@ -18,7 +18,7 @@ public class Charge extends Point {
     /**
      * Set the point's x and y coordinates, and size.
      */
-    public void set(int x, int y, int size) {
+    public void set(int x, int y, double size) {
         set(x, y);
         this.size = size;
     }
